refactor in a module register
