+++ ereuse_devicehub/register/templates/registration/email_admin_new_user.txt
