+++ ereuse_devicehub/register/templates/registration/email_validation.html
@@ -1,0 +1,9 @@
+Welcome to Usody.com, {{ name }}!<br />
+<br />
+Please confirm your account by clicking on the following link:<br />
+<a href="{{ url }}">{{ url }}</a><br />
+<br />
+<img src="https://{{ host }}{{ url_for('static', filename='img/usody_logo_transparent_noicon-y-purple-120x41.png') }}" alt="">
+<br />
+Plaça Eusebi Güell 6-7, Edifici Vèrtex (UPC), planta 0, Barcelona 08034, Spain<br />    
+Associació Pangea – Coordinadora Comunicació per a la Cooperació  - NIF: G-60437761
