+++ ereuse_devicehub/register/templates/registration/email_validation.txt
