+++ ereuse_devicehub/templates/ereuse_devicehub/user_login.html
@@ -61,7 +61,7 @@
                     <button class="btn btn-primary w-100" type="submit">Next</button>
                   </div>
                   <div class="col-12">
-                    <p class="small mb-0">Don't have account? <a href="{{ url_for('core.user-registration') }}">Create an account</a></p>
+                    <p class="small mb-0">Don't have account? <a href="{{ url_register }}">Create an account</a></p>
                   </div>
                 </form>
